Move DataSelector to components/index/cards/charts
