+++ components/index/cards/charts/TimeBarChart.vue
@@ -83,7 +83,7 @@ import { Chart } from 'chart.js'
 import dayjs from 'dayjs'
 import { GraphDataType } from '@/utils/formatGraph'
 import DataView from '@/components/DataView.vue'
-import DataSelector from '@/components/DataSelector.vue'
+import DataSelector from '@/components/index/cards/charts/DataSelector.vue'
 import DataViewBasicInfoPanel from '@/components/index/cards/charts/DataViewBasicInfoPanel.vue'
 import OpenDataLink from '@/components/OpenDataLink.vue'
 import { DisplayData, yAxesBgPlugin, scrollPlugin } from '@/plugins/vue-chart'
