+++ components/index/cards/charts/TimeStackedBarChart.vue
@@ -102,7 +102,7 @@ import { TranslateResult } from 'vue-i18n'
 import { Chart } from 'chart.js'
 import dayjs from 'dayjs'
 import DataView from '@/components/DataView.vue'
-import DataSelector from '@/components/DataSelector.vue'
+import DataSelector from '@/components/index/cards/charts/DataSelector.vue'
 import DataViewBasicInfoPanel from '@/components/index/cards/charts/DataViewBasicInfoPanel.vue'
 import { DisplayData, yAxesBgPlugin, scrollPlugin } from '@/plugins/vue-chart'
 import { getGraphSeriesStyle, SurfaceStyle } from '@/utils/colors'
